working gratitudes, working import and export, and generally overall working

+++ src/store/useDayGratitudes.ts
@@ -1,40 +1,43 @@
 import { StorageServiceKey } from "@/core/constants/injectionKeys";
 import { inject, reactive } from "vue";
-import useDayState from "./useDayState";
 import { IGratitude } from "@/core/interfaces/entities/IGratitude";
 import { LoggingService } from "@/core/services/loggingService";
+import { type DateTime } from "luxon";
 
 export default function useDayGratitudes() {
-	const { selectedDay } = useDayState();
 
 	const storageService = inject(StorageServiceKey)!;
 	const gratitudes = reactive<IGratitude[]>([]);
 
-	async function fetchGratitudes() {
+	async function fetchGratitudes(day: DateTime) {
 		LoggingService.log("fetch gratitudes");
-		const dayGratitudes = await storageService.getGratitudes(selectedDay.value);
-		Object.assign(gratitudes, dayGratitudes);
+        gratitudes.length = 0;
+		Object.assign(gratitudes, await storageService.getGratitudes(day));
 	}
 
 	async function createGratitude(gratitude: IGratitude) {
 		LoggingService.log("create gratitudes", gratitude);
+		gratitude.id = await storageService.createGratitude(gratitude);
 		gratitudes.push(gratitude);
-		await storageService.createGratitude(gratitude);
 	}
 
-	async function updateGratitude(gratitude: IGratitude) {
+	async function updateGratitude(id: number|undefined, gratitude: IGratitude) {
+        if (!id) return;
 		LoggingService.log("update gratitudes", gratitude);
-		await storageService.updateGratitude(gratitude);
+        const oldGratitude = gratitudes.find((g) => g.id === id);
+        if (!oldGratitude) return;
+        Object.assign(oldGratitude, gratitude);
+		await storageService.updateGratitude(id, gratitude);
 	}
 
-	async function deleteGratitude(id: number) {
+	async function deleteGratitude(id: number|undefined) {
+        if (!id) return;
 		LoggingService.log("delete gratitudes", id);
 		const index = gratitudes.findIndex((g) => g.id === id);
 		if (index > -1) gratitudes.splice(index, 1);
 		await storageService.deleteGratitude(id);
 	}
 
-	// watch(selectedDay, () => fetchGratitudes(), { immediate: true });
 
 	return {
 		gratitudes,
